DEEP-204 refactoring PR comments
